fixed edit label styling on label list
