+++ src/components/Settings/DeleteLabel/DeleteLabel.module.scss
@@ -18,6 +18,7 @@
   color: lightgray;
   cursor: pointer;
   pointer-events: auto;
+  margin-top: 3px;
   :hover {
     color: gray;
   }
