+++ src/components/Settings/UpdateLabelModal/UpdateLabelModal.module.scss
@@ -1,0 +1,63 @@
+@import '../../../styles/variables.scss';
+
+.editModalStyle {
+  position: relative !important;
+  height: 225px !important;
+  width: 350px !important;
+}
+
+.labelDesign {
+  display: inline-block;
+  color: white;
+  text-align: center;
+  padding-top: 2px;
+  width: 100px;
+  height: 20px;
+  border-radius: 25px;
+  font-size: 0.8rem;
+}
+
+.hover {
+  cursor: pointer;
+  pointer-events: auto;
+}
+
+.headerDiv {
+  display: flex !important;
+  flex-direction: row !important;
+}
+
+.closeButton {
+  background: none;
+  border: none;
+  color: lightgray;
+  font-size: 1.5rem;
+  font-weight: 100;
+  margin-right: 85px;
+  cursor: pointer;
+  pointer-events: auto;
+  outline: none;
+  &hover {
+    color: gray;
+  }
+}
+
+.editLabelButton {
+  color: #f54987;
+  background: none;
+  border: none;
+  font-size: 1rem;
+  cursor: pointer;
+  pointer-events: auto;
+  outline: none;
+}
+
+.button {
+  background: #f87da9 !important;
+  color: #fee5ee !important;
+}
+
+.button:hover {
+  background: #f54987 !important;
+  color: white !important;
+}
